feat(rating): improve leaderboard layout and profit formatting

+++ app/src/main/java/com/example/traderapp/data/model/UserData.kt
@@ -3,7 +3,7 @@ package com.example.traderapp.data.model
 data class UserData(val email: String = "",
                     val balance: Double = 1000.00,
                     val tradeVolume: Int = 0,
-                    val profit: Int = 0,
+                    val profit: Double = 0.0,
                     val cryptoList: List<CryptoDto> = emptyList(),
                     val initialTotalBalance: Double = 1000.00,
                     )
